refactor: CartListSection 리팩터링
